Add peek function in class BitString
Sometimes the packetlib needs to peek at unparsed bits without actually parsing them. For example, when parsing a GRE header, packetlib needs to peek at the unparsed bits to check if checksum_present bit is set, and then calculate the total size of the GRE header accordingly. This CL adds a peek function that can be used in such cases.

PiperOrigin-RevId: 460375660